requirements.txt
amqp==2.5.2
atomicwrites==1.3.0
attrs==19.3.0
beautifulsoup4==4.8.1
behave==1.2.6
billiard==3.6.3.0
bs4==0.0.1
celery==4.4.2
certifi==2019.9.11
chardet==3.0.4
colorama==0.4.1
coverage==5.0.4
dj-database-url==0.5.0
Django==2.2.7
django-heroku==0.3.1
djangorestframework==3.11.0
djangorestframework-jwt==1.11.0
fact==1.0.0
fake-useragent==0.1.11
Faker==0.9.1
gunicorn==20.0.4
idna==2.8
importlib-metadata==0.23
kombu==4.6.8
lxml==4.4.1
mixer==6.1.3
more-itertools==7.2.0
myMath==1.0.1
packaging==19.2
parse==1.12.1
parse-type==0.5.2
pluggy==0.13.1
psycopg2==2.8.4
py==1.8.0
pygame==1.9.6
PyJWT==1.7.1
pyparsing==2.4.5
pytest==5.3.1
python-dateutil==2.8.1
pytz==2019.3
redis==3.5.2
requests==2.22.0
six==1.13.0
soupsieve==1.9.5
sqlparse==0.3.0
text-unidecode==1.2
urllib3==1.25.7
vine==1.3.0
virtualenv==16.7.7
vk==2.0.2
vk-api==11.8.0
wcwidth==0.1.7
whitenoise==5.0.1
zipp==0.6.0
